requirements.txt
duckduckgo-search 
newspaper4k 
lxml_html_clean 
phidata
uvicorn
fastapi
packaging
groq